+++ server/src-lib/Hasura/GraphQL/Resolve.hs
@@ -45,6 +45,8 @@ data QueryRootFldAST v
   | QRFSimple !(DS.AnnSimpleSelG v)
   | QRFAgg !(DS.AnnAggSelG v)
   | QRFActionSelect !(DS.AnnSimpleSelG v)
+  | QRFActionExecuteObject !(DS.AnnSimpleSelG v)
+  | QRFActionExecuteList !(DS.AnnSimpleSelG v)
   deriving (Show, Eq)
 
 type QueryRootFldUnresolved = QueryRootFldAST UnresolvedVal
@@ -56,17 +58,21 @@ traverseQueryRootFldAST
   -> QueryRootFldAST a
   -> f (QueryRootFldAST b)
 traverseQueryRootFldAST f = \case
-  QRFPk s           -> QRFPk <$> DS.traverseAnnSimpleSel f s
-  QRFSimple s       -> QRFSimple <$> DS.traverseAnnSimpleSel f s
-  QRFAgg s          -> QRFAgg <$> DS.traverseAnnAggSel f s
-  QRFActionSelect s -> QRFActionSelect <$> DS.traverseAnnSimpleSel f s
+  QRFPk s                  -> QRFPk <$> DS.traverseAnnSimpleSel f s
+  QRFSimple s              -> QRFSimple <$> DS.traverseAnnSimpleSel f s
+  QRFAgg s                 -> QRFAgg <$> DS.traverseAnnAggSel f s
+  QRFActionSelect s        -> QRFActionSelect <$> DS.traverseAnnSimpleSel f s
+  QRFActionExecuteObject s -> QRFActionExecuteObject <$> DS.traverseAnnSimpleSel f s
+  QRFActionExecuteList s   -> QRFActionExecuteList <$> DS.traverseAnnSimpleSel f s
 
 toPGQuery :: QueryRootFldResolved -> Q.Query
 toPGQuery = \case
-  QRFPk s           -> DS.selectQuerySQL DS.JASSingleObject s
-  QRFSimple s       -> DS.selectQuerySQL DS.JASMultipleRows s
-  QRFAgg s          -> DS.selectAggQuerySQL s
-  QRFActionSelect s -> DS.selectQuerySQL DS.JASSingleObject s
+  QRFPk s                  -> DS.selectQuerySQL DS.JASSingleObject s
+  QRFSimple s              -> DS.selectQuerySQL DS.JASMultipleRows s
+  QRFAgg s                 -> DS.selectAggQuerySQL s
+  QRFActionSelect s        -> DS.selectQuerySQL DS.JASSingleObject s
+  QRFActionExecuteObject s -> DS.selectQuerySQL DS.JASSingleObject s
+  QRFActionExecuteList s   -> DS.selectQuerySQL DS.JASMultipleRows s
 
 validateHdrs
   :: (Foldable t, QErrM m) => UserInfo -> t Text -> m ()
@@ -77,14 +83,22 @@ validateHdrs userInfo hdrs = do
     throw400 NotFound $ hdr <<> " header is expected but not found"
 
 queryFldToPGAST
-  :: ( MonadReusability m, MonadError QErr m, MonadReader r m, Has FieldMap r
-     , Has OrdByCtx r, Has SQLGenCtx r, Has UserInfo r
+  :: ( MonadReusability m
+     , MonadError QErr m
+     , MonadReader r m
+     , Has FieldMap r
+     , Has OrdByCtx r
+     , Has SQLGenCtx r
+     , Has UserInfo r
      , Has QueryCtxMap r
+     , HasVersion
+     , MonadIO m
      )
-  => G.Field
+  => V.Field
+  -> RA.QueryActionExecuter
   -> m QueryRootFldUnresolved
-queryFldToPGAST fld = do
-  opCtx <- getOpCtx $ G._fName fld
+queryFldToPGAST fld actionExecuter = do
+  opCtx <- getOpCtx $ V._fName fld
   userInfo <- asks getter
   case opCtx of
     QCSelect ctx -> do
@@ -102,8 +116,16 @@ queryFldToPGAST fld = do
     QCFuncAggQuery ctx -> do
       validateHdrs userInfo (_fqocHeaders ctx)
       QRFAgg <$> RS.convertFuncQueryAgg ctx fld
-    QCActionFetch ctx ->
+    QCAsyncActionFetch ctx ->
       QRFActionSelect <$> RA.resolveAsyncActionQuery userInfo ctx fld
+    QCAction ctx -> do
+      case jsonAggType of
+        DS.JASMultipleRows -> QRFActionExecuteList
+        DS.JASSingleObject -> QRFActionExecuteObject
+      <$> actionExecuter (RA.resolveActionQuery fld ctx (userVars userInfo))
+      where
+        outputType = _saecOutputType ctx
+        jsonAggType = RA.mkJsonAggSelect outputType
 
 mutFldToTx
   :: ( HasVersion