+++ server/src-lib/Hasura/GraphQL/Validate/Types.hs
@@ -92,7 +92,7 @@ import qualified Data.HashMap.Strict.InsOrd    as OMap
 import qualified Data.HashSet                  as Set
 import qualified Data.Text                     as T
 import qualified Language.GraphQL.Draft.Syntax as G
-import qualified Language.GraphQL.Draft.TH     as G
+-- import qualified Language.GraphQL.Draft.TH     as G
 import qualified Language.Haskell.TH.Syntax    as TH
 
 import           Control.Lens                  (makePrisms)